Split unit and instrumentation tests
