+++ .github/workflows/test-plugin.yml
@@ -22,7 +22,7 @@ jobs:
           key: gradle-caches-v0-${{ hashFiles('build.gradle.kts', 'settings.gradle.kts', 'gradle/wrapper/gradle-wrapper.properties') }}
 
       - name: Run tests
-        run: ./gradlew test
+        run: ./gradlew test --tests 'com.ioki.lokalise.gradle.plugin.unit*'
 
       - name: Publish Test Report
         uses: mikepenz/action-junit-report@v3
@@ -51,7 +51,7 @@ jobs:
         run: ./gradlew publishToMavenLocal
 
       - name: Run publishing/consuming tests
-        run: ./gradlew test --tests 'LokaliseGradlePluginPublishingTest'
+        run: ./gradlew test --tests 'com.ioki.lokalise.gradle.plugin.instrumentation*'
 
       - name: Publish Test Report
         uses: mikepenz/action-junit-report@v3
